optim

+++ get_next_line/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: yonadry <yonadry@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/28 18:49:34 by yonadry           #+#    #+#             */
-/*   Updated: 2022/11/30 14:17:00 by yonadry          ###   ########.fr       */
+/*   Updated: 2023/02/18 11:56:50 by yonadry          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@
 # include<fcntl.h>
 
 char	*get_next_line(int fd);
-size_t	ft_strlen(char *s);
+int	ft_strlen(char *s);
 int		newline_found(char *read_line);
 char	*ft_strjoin(char *s1, char *s2);
 char	*ft_strdup(char *s1);
